Reverted US_PHONE_NUMBER_REGEXP to its old format (before https://github.com/code-dot-org/code-dot-org/pull/8649), and moved pegasus/test/test_regexp_utils tests into shared/test/test_regexp.
